+++ packages/styleguide/stories/Atoms/Icons/constants.tsx
@@ -11,6 +11,7 @@ export const VENDOR_ICONS = pick(
   [
     'AmexIcon',
     'DiscordIcon',
+    'DiscordOutlineIcon',
     'FacebookIcon',
     'FaviconIcon',
     'FaviconSolidIcon',
@@ -18,8 +19,6 @@ export const VENDOR_ICONS = pick(
     'GithubOutlineIcon',
     'GplusIcon',
     'InstagramIcon',
-    'JavaIcon',
-    'JavascriptIcon',
     'JiraIcon',
     'LinkedinIcon',
     'LinkedinOutlineIcon',
@@ -27,11 +26,9 @@ export const VENDOR_ICONS = pick(
     'MediumIcon',
     'OpenAiFullIcon',
     'OpenAiIcon',
-    'PythonIcon',
     'RailsIcon',
     'ReactIcon',
     'RedditIcon',
-    'RubyIcon',
     'SlackIcon',
     'TikTokIcon',
     'TwitterIcon',
@@ -80,4 +77,45 @@ export const LE_ICONS = pick(
   ].sort()
 );
 
-export const UI_ICONS = omit(ICONS, keys({ ...LE_ICONS, ...VENDOR_ICONS }));
+export const SKILLS_ICONS = pick(
+  ICONS,
+  [
+    'ArtificialIntelligenceIcon',
+    'BashShellIcon',
+    'CIcon',
+    'CPlusIcon',
+    'CSharpIcon',
+    'CloudComputingIcon',
+    'ComputerScienceIcon',
+    'CybersecurityIcon',
+    'DataAnalyticsIcon',
+    'DataEngineeringIcon',
+    'DataScienceIcon',
+    'DataVisualizationIcon',
+    'DevopsIcon',
+    'FallbackSkillIcon',
+    'GameDevelopmentIcon',
+    'GoIcon',
+    'HtmlCssIcon',
+    'ItIcon',
+    'JavaIcon',
+    'JavascriptIcon',
+    'KotlinIcon',
+    'MachineLearningIcon',
+    'MathIcon',
+    'MobileDevelopmentIcon',
+    'PhpIcon',
+    'PythonIcon',
+    'RIcon',
+    'RubyIcon',
+    'SqlIcon',
+    'SwiftIcon',
+    'WebDesignIcon',
+    'WebDevelopmentIcon',
+  ].sort()
+);
+
+export const UI_ICONS = omit(
+  ICONS,
+  keys({ ...LE_ICONS, ...VENDOR_ICONS, ...SKILLS_ICONS })
+);